+++ R/baguette-package.R
@@ -24,10 +24,6 @@
 #' @keywords internal
 "_PACKAGE"
 
-#' @importFrom magrittr %>%
-#' @export
-magrittr::`%>%`
-
 #' @importFrom generics var_imp
 #' @export
 generics::var_imp