+++ R/mars.R
@@ -8,7 +8,7 @@ mars_bagger <- function(rs, control, ..., call) {
   iter <- get_iterator(control)
 
   rs <-
-    rs %>%
+    rs |>
     dplyr::mutate(model = iter(
       fit_seed,
       splits,
@@ -27,12 +27,12 @@ mars_bagger <- function(rs, control, ..., call) {
   imps <- compute_imp(rs, mars_imp, control$var_imp)
 
   rs <-
-    rs %>%
+    rs |>
     replace_parsnip_terms()
 
   if (control$reduce) {
     rs <-
-     rs %>%
+     rs |>
       dplyr::mutate(model = purrr::map(model, axe_mars))
   }