+++ tests/testthat/test-cart.R
@@ -104,9 +104,9 @@ test_that('check CART parsnip interface', {
 
   set.seed(4779)
   expect_error(
-    reg_mod <- bag_tree(cost_complexity = .001, min_n = 3) %>%
-      set_engine("rpart", times = 3) %>%
-      set_mode("regression") %>%
+    reg_mod <- bag_tree(cost_complexity = .001, min_n = 3) |>
+      set_engine("rpart", times = 3) |>
+      set_mode("regression") |>
       fit(mpg ~ ., data = mtcars),
     regexp = NA
   )
@@ -126,9 +126,9 @@ test_that('check CART parsnip interface', {
 
   set.seed(4779)
   expect_error(
-    class_cost <- bag_tree(min_n = 3, class_cost = 2) %>%
-      set_engine("rpart", times = 3) %>%
-      set_mode("classification") %>%
+    class_cost <- bag_tree(min_n = 3, class_cost = 2) |>
+      set_engine("rpart", times = 3) |>
+      set_mode("classification") |>
       fit(Class ~ ., data = two_class_dat),
     regexp = NA
   )
@@ -158,9 +158,9 @@ test_that('check CART parsnip interface', {
 
   set.seed(4779)
   expect_error(
-    class_mod <- bag_tree(cost_complexity = .001, min_n = 3) %>%
-      set_engine("rpart", times = 3) %>%
-      set_mode("classification") %>%
+    class_mod <- bag_tree(cost_complexity = .001, min_n = 3) |>
+      set_engine("rpart", times = 3) |>
+      set_mode("classification") |>
       fit(Class ~ ., data = two_class_dat),
     regexp = NA
   )
@@ -189,15 +189,15 @@ test_that('check CART parsnip interface', {
 
 test_that('mode specific package dependencies', {
   expect_identical(
-    get_from_env(paste0("bag_tree", "_pkgs")) %>%
-      dplyr::filter(engine == "rpart", mode == "classification") %>%
+    get_from_env(paste0("bag_tree", "_pkgs")) |>
+      dplyr::filter(engine == "rpart", mode == "classification") |>
       dplyr::pull(pkg),
     list(c("rpart", "baguette"))
   )
 
   expect_identical(
-    get_from_env(paste0("bag_tree", "_pkgs")) %>%
-      dplyr::filter(engine == "rpart", mode == "regression") %>%
+    get_from_env(paste0("bag_tree", "_pkgs")) |>
+      dplyr::filter(engine == "rpart", mode == "regression") |>
       dplyr::pull(pkg),
     list(c("rpart", "baguette"))
   )